+++ app/js/ttth/crashReporting.js
@@ -27,11 +27,25 @@ Sentry.init({
 // simple way to force a crash:
 // myUndefinedFunction();
 
+function enableSentry () {
+    Sentry.getCurrentHub().getClient().getOptions().enabled = true
+    console.log('Sentry is now enabled')
+}
+
+function disableSentry () {
+    Sentry.getCurrentHub().getClient().getOptions().enabled = false
+    console.warn('Sentry is now disabled')
+}
+
+// export both functions
+// exports.enableSentry = enableSentry
+// exports.disableSentry = disableSentry
+
 // ----------------------------------------------------------------------------
 // Error Handling using: electron-unhandled (https://github.com/sindresorhus/electron-unhandled)
 // ----------------------------------------------------------------------------
 const unhandled = require('electron-unhandled')
-const {openNewGitHubIssue, debugInfo} = require('electron-util')
+const { openNewGitHubIssue, debugInfo } = require('electron-util')
 
 unhandled({
     showDialog: true,
@@ -40,6 +54,6 @@ unhandled({
             user: 'yafp',
             repo: 'ttth',
             body: `\`\`\`\n${error.stack}\n\`\`\`\n\n---\n\n${debugInfo()}`
-        });
+        })
     }
-});
+})